Updating for custom background image paths on linux/ios/android
