+++ carveracontroller/main.py
@@ -66,6 +66,9 @@ class Lang(Observable):
 import json
 import re
 import tempfile
+import os
+import platform
+import subprocess
 from kivy.app import App
 from kivy.clock import Clock
 from kivy.uix.widget import Widget
@@ -427,9 +430,13 @@ class CoordPopup(ModalView):
         self.MoveA_popup = MoveAPopup(self)
         self.mode = 'Run' # 'Margin' / 'ZProbe' / 'Leveling'
         super(CoordPopup, self).__init__(**kwargs)
+        default_bkg_images = os.path.join(os.path.dirname(__file__), 'data/play_file_image_backgrounds')
         self.background_image_files = [
-            f.replace(".png", "") for f in os.listdir(PLAY_FILE_IMAGE_DIR) if f.endswith(".png")
+            f.replace(".png", "") for f in os.listdir(PLAY_FILE_IMAGE_DIR) if f.endswith(".png")            
         ]
+        for f in os.listdir(default_bkg_images):
+            if f.endswith(".png"):
+                self.background_image_files.append(f.replace(".png", ""))
 
 
         # Ensure the spinner is updated after initialization
@@ -441,13 +448,50 @@ class CoordPopup(ModalView):
 
     def update_background_image(self, filename):
         if filename != "None":
+            old_source = os.path.join(os.path.dirname(__file__), 'data/play_file_image_backgrounds', filename)
             new_source = os.path.join(PLAY_FILE_IMAGE_DIR, filename)
             cnc_workspace = self.ids.cnc_workspace
-            cnc_workspace.update_background_image(new_source + ".png")
+            if os.path.isfile(new_source + ".png"):
+                cnc_workspace.update_background_image(new_source + ".png")
+            elif os.path.isfile(old_source + ".png"):
+                cnc_workspace.update_background_image(old_source + ".png")
+            else:
+                cnc_workspace.update_background_image("None")
         else:
             cnc_workspace = self.ids.cnc_workspace
             cnc_workspace.update_background_image("None")
     
+    def open_bkg_img_dir(self):
+        folder_path = PLAY_FILE_IMAGE_DIR
+
+        # Ensure the folder exists
+        if not os.path.exists(folder_path):
+            print(f"Folder '{folder_path}' does not exist!")
+            return
+
+        # Open based on OS
+        if platform.system() == "Windows":
+            os.startfile(folder_path)
+        elif platform.system() == "Darwin":  # macOS
+            subprocess.Popen(["open", folder_path])
+        else:  # Linux
+            subprocess.Popen(["xdg-open", folder_path])
+        
+        folder_path = os.path.join(os.path.dirname(__file__), 'data/play_file_image_backgrounds')
+
+        # Ensure the folder exists
+        if not os.path.exists(folder_path):
+            print(f"Folder '{folder_path}' does not exist!")
+            return
+
+        # Open based on OS
+        if platform.system() == "Windows":
+            os.startfile(folder_path)
+        elif platform.system() == "Darwin":  # macOS
+            subprocess.Popen(["open", folder_path])
+        else:  # Linux
+            subprocess.Popen(["xdg-open", folder_path])
+
     def set_config(self, key1, key2, value):
         self.config[key1][key2] = value
         self.cnc_workspace.draw()
@@ -3870,6 +3914,8 @@ def android_tweaks():
 def load_app_configs():
     if Config.has_option('carvera', 'ui_density_override') and Config.get('carvera', 'ui_density_override') == "1":
         Metrics.set_density(float(Config.get('carvera', 'ui_density')))
+    if Config.has_option('carvera', 'custom_bkg_img_dir'):
+        PLAY_FILE_IMAGE_DIR = Config.get('carvera', 'custom_bkg_img_dir')
 
 def set_config_defaults(default_lang):
     if not Config.has_section('carvera'):
@@ -3937,7 +3983,7 @@ def load_constants():
     SHORT_LOAD_TIMEOUT = 3  # s
     WIFI_LOAD_TIMEOUT = 30 # s
     HEARTBEAT_TIMEOUT = 10
-    PLAY_FILE_IMAGE_DIR = os.path.join(os.path.dirname(__file__), 'data/play_file_image_backgrounds')
+    PLAY_FILE_IMAGE_DIR = Config.get('carvera', 'custom_bkg_img_dir')
     MAX_TOUCH_INTERVAL = 0.15
     GCODE_VIEW_SPEED = 1
